changed receiverId to big int #fix #story[1126]

+++ server/src/main/resources/reset.sql
@@ -91,7 +91,7 @@ CREATE TABLE IF NOT EXISTS `TRANSFERS` (
   `StartLon` double NOT NULL,
   `EndLat` double NOT NULL,
   `EndLon` double NOT NULL,
-  `ReceiverId` int(11) NOT NULL,
+  `ReceiverId` bigint(11) NOT NULL,
   `ArrivalTime` timestamp NOT NULL,
   `OrganType` text NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
